+++ example/text_form_builder_example/android/app/src/main/kotlin/com/example/text_form_builder_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.text_form_builder_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}