+++ lib/errors/text_form_builder_error.dart
@@ -1,0 +1,9 @@
+part of '../text_form_builder.dart';
+
+class TextFormBuilderError implements Exception {
+  final String message;
+
+  TextFormBuilderError([this.message = "Unknown TextFormBuilderError"]);
+
+  String toString() => message;
+}