+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## [0.0.1] - TODO: Add release date.
+# Change Log
 
-* TODO: Describe initial release.
+## [0.0.1] - 29.09.2020
+
+### Added
+* Generate a form with its the boilerplate code
+* Add text fields aswell as other widgets to your generated form
+* Disable that submitting the last field in the form submits the form itself
+* Disable that clicking outside of a field closes the keyboard
+
+### Changed
+* Nothing
+
+### Fixed
+* Nothing